finished the add a valorant item to roomdb

+++ data/src/main/java/com/serdar/data/dto/weapons/detail/WeaponStats.kt
@@ -9,7 +9,7 @@ data class WeaponStats(
     val equipTimeSeconds: Double,
     val feature: String,
     val fireMode: Any,
-    val fireRate: Int,
+    val fireRate: Double,
     val firstBulletAccuracy: Double,
     val magazineSize: Int,
     val reloadTimeSeconds: Int,
